As described in #1484 comment https://github.com/dtcenter/MET/issues/1484#issuecomment-699521164, port bugfix from PR #1486 into the main_v9.1 branch.
